FER_2020_python/requirements.txt
flask
flask_restful
werkzeug
gunicorn
numpy==1.18.5
requests
keras.models
tensorflow